backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.1
python-dotenv==1.0.0
google-generativeai==0.3.1
openai-whisper==20231117
numpy==1.26.2
torch==2.1.1
torchaudio==2.1.1

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
coverage==7.3.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 